Added Functionality
I added what i think will basically play the game. With this setup if it
was moved to hardware the game would actually play as the computer
reacts everytime that user input is changed. Included a text of the
compilation erros of the latest file (most of it is coming from Dave's
code but it could be caused by code i added)